requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==1.5.7
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-cors-headers==3.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.7.0
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytz==2021.1
redis==6.4.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.4.2
toml==0.10.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.10.0
zope.interface==7.2
